+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+## Changed
+* [breaking] Handlers are no longer stored internally. The `poll()` function is now required to
+perform the mapping to individual handlers. This is intended to make the handler API signature more
+flexible, as not all handlers may need the same context.
+
 ## [0.4.0] - 2024-06-13
 
 * [breaking] Updated to `minimq` v0.9.0